Bugfix
X信号出力を保持する変数 matrix が uint8_t になっていたバグを修正

+++ usb_keyboard_bridge_for_msx/usb_keyboard_bridge_for_msx.c
@@ -402,7 +402,7 @@ static void initialization( void ) {
 
 // --------------------------------------------------------------------
 void response_core( void ) {
-	uint8_t matrix;
+	uint32_t matrix;
 	int y;
 	static const uint32_t x_mask = 0x0FF << MSX_KEYMATRIX_RESULT_PIN;
 	#if DEBUG_UART_ON
